requirements.txt
requests>=2.25.1
Pillow>=8.0.0
numpy>=1.19.0
torch>=1.7.0 